Add files via upload
It’s a web page using HTML, CSS and JavaScript/JQuery. 
JavaScript, Event Handling, Bindings and DOM.
CRUD Operation are performed using HTML, CSS and JavaScript/JQuery.
When user fill the form and click 'add' button, the record will be added in the table on the left side of page.
“Action” column contains Update and Remove links.
"Update" loads the record in the form on right. 
When user update the values and press 'update' button, it will update the corresponding row in the table on left.
Initially, when the page is loaded, the table will be empty and "Update" button will be disabled. 
When user click the "Update" link shown under 'Action' column, "add" button will be disabled. 
The "reset" button shall re-initialize the form at the same state when the page was loaded to add new records, so it will make the add button enabled. 
The page will not reload during these operations, because it’s a single page app.